web/requirements.txt
Flask==3.0.0
pyserial==3.5
ansi2html==1.9.1